% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_ementa.R
\name{gemini_ementa}
\alias{gemini_ementa}
\title{Processa ementas utilizando o Gemini e salva os resultados em arquivos JSON}
\usage{
gemini_ementa(
  ementa,
  identificador = NULL,
  instrucoes,
  perguntas,
  chaves,
  api_key = NULL,
  temperatura = 0.2,
  diretorio = "."
)
}
\arguments{
\item{ementa}{Lista de ementas a serem processadas.}

\item{identificador}{Vetor opcional de identificadores para as ementas. Se não fornecido, será gerado automaticamente.}

\item{instrucoes}{Instruções específicas para o processamento das ementas.}

\item{perguntas}{Perguntas a serem respondidas com base nas ementas.}

\item{chaves}{Palavras-chave para a extração de informações.}

\item{api_key}{Chave da API do serviço Gemini. Se não fornecida, a função tentará usar a variável de ambiente GEMINI_KEY.}

\item{temperatura}{Parâmetro de temperatura para ajustar a aleatoriedade das respostas da IA. O valor padrão é 0.2.}

\item{diretorio}{Diretório onde os arquivos JSON serão salvos. O padrão é o diretório atual.}
}
\value{
Nenhum valor é retornado. Os resultados são salvos em arquivos JSON no diretório especificado.
}
\description{
Esta função processa uma lista de ementas utilizando o serviço Gemini, extrai informações conforme instruções fornecidas,
responde a perguntas específicas e salva os resultados em arquivos JSON. Se um identificador não for fornecido, ele será gerado automaticamente.
A função verifica e evita duplicatas baseando-se em arquivos existentes no diretório especificado.
}
\examples{
\dontrun{
ementas <- c("Ementa 1...", "Ementa 2...")
instrucoes <- "Instruções gerais"
perguntas <- c("Pergunta 1", "Pergunta 2")
chaves <- c("Chave 1", "Chave 2")

gemini_ementa(ementas, 
              instrucoes = "Instruções específicas", 
              perguntas = "Perguntas específicas", 
              chaves = "Palavras-chave",
              api_key = "SUA_API_KEY")
}
}
