% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_ler.R
\name{gemini_ler}
\alias{gemini_ler}
\title{Lê e Combina Arquivos JSON}
\usage{
gemini_ler(arquivos, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de strings. Caminhos para os arquivos JSON a serem lidos. Se NULL, todos os arquivos JSON no diretório serão lidos.}

\item{diretorio}{String. Diretório onde os arquivos JSON estão localizados. O padrão é o diretório atual.}
}
\value{
Um data frame contendo os dados combinados de todos os arquivos JSON.
}
\description{
Esta função lê múltiplos arquivos JSON de um diretório especificado e os combina em um único data frame.
}
\examples{
\dontrun{
# Ler todos os arquivos JSON no diretório atual
df <- gemini_ler(diretorio = ".")

# Ler arquivos JSON específicos
df <- gemini_ler(arquivos = c("file1.json", "file2.json"))
}

}
