% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World}
\usage{
hello(name)
}
\arguments{
\item{name}{}
}
\value{
A semple menssage
}
\description{
This is a simple function to test the package creation process.
}
\examples{
Hello World!
}
